+++ finalynx/portfolio/node.py
@@ -10,12 +10,15 @@ from rich.tree import Tree
 
 from ..config import DEFAULT_CURRENCY
 from ..config import get_active_theme as TH
+from .constants import AssetClass
+from .constants import AssetSubclass
 from .constants import LinePerf
 from .hierarchy import Hierarchy
 from .render import Render
 from .targets import Target
 
 if TYPE_CHECKING:
+    from .envelope import Envelope
     from .folder import Folder
 
 
@@ -25,12 +28,16 @@ class Node(Hierarchy, Render):
     def __init__(
         self,
         name: str,
+        asset_class: AssetClass = AssetClass.UNKNOWN,
+        asset_subclass: AssetSubclass = AssetSubclass.UNKNOWN,
         parent: Optional["Folder"] = None,
         target: Optional[Target] = None,
         newline: bool = False,
+        perf: Optional[LinePerf] = None,
+        currency: Optional[str] = None,
+        envelope: Optional["Envelope"] = None,
         aliases: Optional[Dict[str, str]] = None,
         agents: Optional[Dict[str, Callable[..., str]]] = None,
-        currency: Optional[str] = None,
     ):
         """This is an abstract class used by the `Line` and `Folder` subclasses.
 
@@ -49,6 +56,10 @@ class Node(Hierarchy, Render):
         self.target = target if target is not None else Target()
         self.target.set_parent(self)
         self.currency = currency if currency else DEFAULT_CURRENCY
+        self.asset_class = asset_class
+        self.asset_subclass = asset_subclass
+        self.perf = perf
+        self.envelope = envelope
 
         if target is not None:
             target.set_parent(self)
@@ -172,7 +183,7 @@ class Node(Hierarchy, Render):
 
     def _render_ideal(self) -> str:
         """:returns: A formatted rendering of the ideal amount to be invested based on the target."""
-        return f"[{TH().ACCENT}]{self.target.render_ideal()}"
+        return f"[{TH().ACCENT}]{self.target.render_ideal()}[/]"
 
     def _render_delta(self, align: bool = True, children: Optional[List["Node"]] = None) -> str:
         """Creates a formatted rendering of the delta investment needed to reach the target.
@@ -180,17 +191,20 @@ class Node(Hierarchy, Render):
         :param children: List of `Node` objects used for vertical alignemnt, defaults to this parent's children.
         :returns: The rendered string.
         """
-        delta, check = round(self.get_delta()), self.target.check()
-        if delta == 0 or check == Target.RESULT_NONE:
+        delta, check, max_length = round(self.get_delta()), self.target.check(), 0
+
+        # Skip nodes where everything is fine, or if the node is empty
+        if delta == 0 or check in [Target.RESULT_NONE, Target.RESULT_OK]:
             return ""
-        color = TH().DELTA_POS if delta > 0 else TH().DELTA_NEG
-        children = children if children else (self.parent.children if self.parent and self.parent.children else [])
-        max_length = np.max([len(str(abs(round(c.get_delta())))) for c in children]) if children else 0
-        max_length = max_length if align else 0
-        if check == Target.RESULT_OK:
-            return f"[{TH().DELTA_OK}]{'✓':>{max_length+3}}[/] "
+
+        # Align all amounts vertically if needed
+        if align:
+            children = children if children else (self.parent.children if self.parent and self.parent.children else [])
+            max_length = np.max([len(str(abs(round(c.get_delta())))) for c in children]) if children else 0
+
         return (
-            f"[{color}]{'+' if delta > 0 else '-'}{abs(delta):>{max_length}} {self._render_currency()}[/] "
+            f"[{TH().DELTA_POS if delta > 0 else TH().DELTA_NEG}]"
+            f"{'+' if delta > 0 else '-'}{abs(delta):>{max_length}} {self._render_currency()}[/] "
             # f"[dim white]→  {self.get_ideal():>{max_length}} {self._render_currency()}[/] "
         )